chore: move GetBooksAndRelatedTasksService to its own file
