+++ __tests__/unit/domain/services/get-books-and-related-tasks-service.spec.js
@@ -1,22 +1,6 @@
+const { GetBooksAndRelatedTasksService } = require('../../../../src/domain/services');
 const { MissingParamError } = require('../../../../src/utils/errors');
 const { DataFakerHelper } = require('../../../helpers');
-
-class GetBooksAndRelatedTasksService {
-  constructor({ getBooksAndRelatedTasksByUserIdRepository, booksAndRelatedTasksSerializer } = {}) {
-    this.getBooksAndRelatedTasksByUserIdRepository = getBooksAndRelatedTasksByUserIdRepository;
-    this.booksAndRelatedTasksSerializer = booksAndRelatedTasksSerializer;
-  }
-
-  async handler({ userId }) {
-    if (!userId) throw new MissingParamError('userId');
-
-    const booksAndRelatedTasks = await this.getBooksAndRelatedTasksByUserIdRepository.get({ userId });
-    if (!booksAndRelatedTasks) return null;
-
-    const booksAndRelatedTasksModel = this.booksAndRelatedTasksSerializer.serialize(booksAndRelatedTasks);
-    return booksAndRelatedTasksModel;
-  }
-}
 
 const makeGetBooksAndRelatedTasksByUserIdRepositorySpy = () => {
   class GetBooksAndRelatedTasksByUserIdRepositorySpy {
