+++ src/domain/services/index.js
@@ -1,7 +1,9 @@
 const CheckIfTokenUserIdIsEqualToProvidedService = require('./check-if-token-user-id-is-equal-to-provided-service');
+const GetBooksAndRelatedTasksService = require('./get-books-and-related-tasks-service');
 const LoginService = require('./login-service');
 
 module.exports = {
   CheckIfTokenUserIdIsEqualToProvidedService,
+  GetBooksAndRelatedTasksService,
   LoginService
 };
